+++ verifiable_planning/verifiable_planning/adapters/__init__.py
@@ -6,4 +6,5 @@ Install/use explicitly, e.g.::
     from verifiable_planning.adapters.llm_planner import plan_from_goal
     from verifiable_planning.adapters.runtime_verify import verify_trace, linear_trace
     from verifiable_planning.adapters.pddl_bridge import plan_to_pddl
+    from verifiable_planning.adapters.planner_bridge import check_plan_with_planner
 """